Modularize too big server
