+++ src/server/server_config.rs
@@ -1,0 +1,63 @@
+use std::collections::HashMap;
+
+use serde::{Deserialize, Serialize};
+
+use crate::shaders::{
+    shader::{GlslSpirvVersion, GlslTargetClient, HlslShaderModel, HlslVersion},
+    shader_error::ShaderErrorSeverity,
+    validator::validator::ValidationParams,
+};
+
+#[allow(non_snake_case)]
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+pub struct ServerHlslConfig {
+    pub shaderModel: HlslShaderModel,
+    pub version: HlslVersion,
+    pub enable16bitTypes: bool,
+}
+#[allow(non_snake_case)]
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+pub struct ServerGlslConfig {
+    pub targetClient: GlslTargetClient,
+    pub spirvVersion: GlslSpirvVersion,
+}
+
+#[allow(non_snake_case)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ServerConfig {
+    pub includes: Vec<String>,
+    pub defines: HashMap<String, String>,
+    pub validate: bool,
+    pub symbols: bool,
+    pub severity: String,
+    pub hlsl: ServerHlslConfig,
+    pub glsl: ServerGlslConfig,
+}
+
+impl ServerConfig {
+    pub fn into_validation_params(&self) -> ValidationParams {
+        ValidationParams {
+            includes: self.includes.clone(),
+            defines: self.defines.clone(),
+            hlsl_shader_model: self.hlsl.shaderModel,
+            hlsl_version: self.hlsl.version,
+            hlsl_enable16bit_types: self.hlsl.enable16bitTypes,
+            glsl_client: self.glsl.targetClient,
+            glsl_spirv: self.glsl.spirvVersion,
+        }
+    }
+}
+
+impl Default for ServerConfig {
+    fn default() -> Self {
+        Self {
+            includes: Vec::new(),
+            defines: HashMap::new(),
+            validate: true,
+            symbols: true,
+            severity: ShaderErrorSeverity::Hint.to_string(),
+            hlsl: ServerHlslConfig::default(),
+            glsl: ServerGlslConfig::default(),
+        }
+    }
+}
